[serde generate] split test file generation.rs by language
